… format/src/test/scala/com/github/sadikovi/riff/MetadataSuite.scala
… format/src/test/scala/com/github/sadikovi/riff/RiffSuite.scala