This PR updates initialization of file reader and writer to remove recursive calls and improve overall performance. As per #54, GC situation has improved and now is comparable with Parquet. I will post benchmarks later, but query update looks like this:
Before patch:
```
Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL Query (int filter, one record):      Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Parquet                                        230 /  272          4.3         230.1       1.0X
ORC                                            757 /  817          1.3         757.4       0.3X
Riff                                           511 /  867          2.0         510.8       0.5X


Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL Query (string filter, one record):   Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Parquet                                        456 /  528          2.2         455.7       1.0X
ORC                                            753 /  910          1.3         753.4       0.6X
Riff                                           361 /  402          2.8         360.7       1.3X
```

After patch:
```
Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL Query (int filter, one record):      Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Parquet                                        231 /  251          4.3         231.1       1.0X
ORC                                            745 /  789          1.3         745.5       0.3X
Riff                                            67 /   76         14.9          67.2       3.4X

Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL Query (string filter, one record):   Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Parquet                                        349 /  375          2.9         349.4       1.0X
ORC                                            702 /  735          1.4         702.3       0.5X
Riff                                            66 /   71         15.2          65.7       5.3X
```

PR also changes Riff API for creating reader/writer/metadata writer/metadata reader.
Benchmarks are updated in gist.